pyproject.toml
[project]
name = "cyclone-impact-toolkit"
version = "1.0.0"
description = "Tropical Cyclone Impact Assessment Toolkit for Fishing Activities - Unified Dashboard"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "beautifulsoup4==4.13.4",
    "contextily==1.6.2",
    "geojson==3.2.0",
    "geopandas==0.14.4",
    "geopy==2.4.1",
    "imageio==2.37",
    "matplotlib==3.7.1",
    "numpy==1.26.4",
    "pandas>=2.3.2",
    "pillow==11.2.1",
    "python-dotenv>=1.2.1",
    "pywebview>=6.0",
    "requests==2.32.3",
    "scipy==1.13.1",
    "shapely==2.1.1",
    "tinydb>=4.8.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pyinstaller>=6.16.0",
    "black>=24.3.0",
    "ruff>=0.3.4",
    "bandit[toml]>=1.7.8",
    "mypy>=1.18.2",
]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["backend", "frontend"]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build"]
skips = ["B101", "B601"]
